Project class should not contain any GUI elements. Status icons moved to main GUI class.
